chore(errors): move the update-error enum to models and lock the TextStyle win
Two ratchets caught this branch, both correctly.

`check_untested_services_floor` flagged `video_metadata_update_error.dart` as a
new service with no test. It is not a service — it is a value type, and it
belongs beside `camera_pick_error.dart` in `lib/models/`, which is where the
sibling enum from this same PR already lives. Moved rather than given a
placeholder test for an enum with no behaviour.

`check_raw_textstyle_ceiling` flagged `explore_view.dart` dropping to zero,
which is a win this PR earned by replacing a raw `TextStyle` with a
`VineTheme` font while rewriting that error branch. `explore_lists_tab.dart`
went 5 -> 4 for the same reason.

The baseline is hand-edited to those two files rather than regenerated.
Regenerating also absorbed drops in `invite_gate_screen`,
`profile_setup_listeners`, `composable_video_grid` and `video_feed_item` —
wins that arrived with the rebase from other people's commits, and are not
this PR's to claim.

Refs #3589
