+++ mobile/lib/widgets/video_metadata/modes/edit/video_metadata_edit_bottom_bar.dart
@@ -11,11 +11,11 @@ import 'package:models/models.dart' show VideoEvent;
 import 'package:openvine/blocs/owner_video_actions/owner_video_actions_cubit.dart';
 import 'package:openvine/l10n/l10n.dart';
 import 'package:openvine/l10n/video_metadata_update_error_l10n.dart';
+import 'package:openvine/models/video_metadata_update_error.dart';
 import 'package:openvine/providers/app_providers.dart';
 import 'package:openvine/providers/creator_delete_enforcement_providers.dart';
 import 'package:openvine/providers/video_editor_provider.dart';
 import 'package:openvine/screens/subtitle_editor/subtitle_editor_screen.dart';
-import 'package:openvine/services/video_metadata_update_error.dart';
 import 'package:openvine/services/video_metadata_update_service.dart';
 import 'package:openvine/utils/delete_result_localization.dart';
 import 'package:openvine/utils/owner_video_cleanup_feedback.dart';
