+++ mobile/scripts/baseline/raw_textstyle.txt
@@ -22,8 +22,7 @@ lib/screens/comments/widgets/comment_item.dart	1
 lib/screens/comments/widgets/comments_empty_state.dart	2
 lib/screens/curated_list_feed_screen.dart	12
 lib/screens/discover_lists_screen.dart	13
-lib/screens/explore/explore_view.dart	1
-lib/screens/explore/tabs/explore_lists_tab.dart	5
+lib/screens/explore/tabs/explore_lists_tab.dart	4
 lib/screens/explore/widgets/explore_buffered_videos_banner.dart	1
 lib/screens/explore_screen_router.dart	1
 lib/screens/feed/video_feed_page/feed_empty_widget.dart	1
