+++ mobile/test/l10n/video_metadata_update_error_l10n_test.dart
@@ -5,7 +5,7 @@ import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:openvine/l10n/generated/app_localizations.dart';
 import 'package:openvine/l10n/video_metadata_update_error_l10n.dart';
-import 'package:openvine/services/video_metadata_update_error.dart';
+import 'package:openvine/models/video_metadata_update_error.dart';
 
 void main() {
   final l10n = lookupAppLocalizations(const Locale('en'));
